requirements.txt
numpy
scipy
matplotlib
tk
mysql-connector-python
python-dotenv